pyproject.toml
[project]
name = "hr5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain-google-genai>=2.0.10",
    "python-dotenv>=1.1.1",
]
